+++ OpenGLCourseAPP/OpenGLCourseAPP/Light.h
@@ -10,8 +10,8 @@ public:
 	Light(GLfloat red, GLfloat green, GLfloat blue, GLfloat aIntensity,
 		GLfloat xDir, GLfloat yDir, GLfloat zDir, GLfloat dIntensity);
 
-	void useLight(GLfloat ambientIntensityLocation, GLfloat ambientColorLocation,
-		GLfloat diffuseIntensityLocation, GLfloat directionLocation);
+	void useLight(GLuint ambientIntensityLocation, GLfloat ambientColorLocation,
+		GLuint diffuseIntensityLocation, GLfloat directionLocation);
 
 	~Light();
